backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.*
pydantic-settings==2.*
openai==1.*
Pillow==11.*
python-dotenv==1.*
python-multipart==0.0.*
aiofiles==24.*
azure-identity==1.*
